+++ src/app/(with-navbar)/signup/view/SectionTitle.tsx
@@ -1,0 +1,17 @@
+import { PropsWithChildren } from "react";
+import { Typography } from "@mui/material";
+import { Title } from "@/app/(with-navbar)/signup/styles";
+
+export default function SectionTitle({ children }: PropsWithChildren) {
+  return (
+    <Title>
+      <Typography
+        variant="h5"
+        component="h1"
+        sx={{ mt: "25px", mb: "20px", fontWeight: 700, whiteSpace: "pre-line" }}
+      >
+        {children}
+      </Typography>
+    </Title>
+  );
+}
